Create folder bin for visualization examples
