+++ visualization/CMakeLists.txt
@@ -1,6 +1,7 @@
 cmake_minimum_required(VERSION 3.16)
 
 add_subdirectory(src)
+add_subdirectory(bin)
 
 if(BUILD_TESTING) 
     add_subdirectory(tests)
