+++ visualization/bin/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_executable("svg-example"
+    svg-example.cpp)
+
+target_link_libraries("svg-example"
+    PRIVATE "visualization")
+
+target_compile_features("svg-example" PRIVATE cxx_std_17)
